Improved code style, redesigned object interaction.
All var changed to const and let. Created a literal object with all numeric constants. Moved player methods to prototype. Removed for loop. The player is passed to the method as an argument. To implement the requirement to separate the update and collision check functions, we had to change engen.js again, so it is also applied for verification. Now after the function of updating all entities is completed, the function of checking for collisions is executed. If a collision is detected, a method is called that sends the player to their original position. Fixed a bug due to which the player had the opportunity to go beyond the playing field below.